requirements.txt
Flask
db-sqlite3
Flask-SQLAlchemy
flask-smorest
